Refactor: Separate responsibilities between controller and RestController
