+++ backend/src/main/java/com/dsapkl/backend/controller/MemberController.java
@@ -8,6 +8,7 @@ import com.dsapkl.backend.entity.Address;
 import com.dsapkl.backend.entity.Member;
 import com.dsapkl.backend.service.CartService;
 import com.dsapkl.backend.service.MemberService;
+import com.dsapkl.backend.util.SessionConst;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
