+++ backend/src/main/java/com/dsapkl/backend/controller/user/CartController.java
@@ -1,4 +1,4 @@
-package com.dsapkl.backend.controller;
+package com.dsapkl.backend.controller.user;
 
 import com.dsapkl.backend.dto.CartForm;
 import com.dsapkl.backend.dto.CartItemForm;
@@ -7,6 +7,7 @@ import com.dsapkl.backend.entity.Member;
 import com.dsapkl.backend.repository.query.CartQueryDto;
 import com.dsapkl.backend.service.CartService;
 import com.dsapkl.backend.service.OrderService;
+import com.dsapkl.backend.util.SessionUtil;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -14,7 +15,6 @@ import com.stripe.Stripe;
 import com.stripe.exception.StripeException;
 import com.stripe.model.checkout.Session;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.*;
@@ -28,6 +28,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Controller
+@RequestMapping("/user")
 @RequiredArgsConstructor
 @Slf4j
 public class CartController {
@@ -57,7 +58,7 @@ public class CartController {
     throws JsonProcessingException {
         Stripe.apiKey = "sk_test_51QclmbPPwZvRdRPfWv7wXxklQBavqLzNsxg3hsnaErdkjaZSvWCncfJXaQ9yUbvxCaUPRfEMsp2GXGwvSd2QHcHn00XH6z4sld";
 
-        Member member = getMember(request);
+        Member member = SessionUtil.getMember(request);
 
         List<CartQueryDto> cartItemListForm = Collections.emptyList();
 
@@ -99,53 +100,6 @@ public class CartController {
         }
 
         return cartItemListForm;
-    }
-
-    /**
-     *  장바구니 담기
-     */
-    @PostMapping("/cart")
-    @ResponseBody
-    public ResponseEntity<String> addCart(@ModelAttribute CartForm cartForm, HttpServletRequest request) {
-
-        Member member = getMember(request);
-        //비로그인 회원은 장바구니를 가질 수 없다.
-        if (member == null) {
-            return new ResponseEntity<String>("로그인이 필요합니다.", HttpStatus.BAD_REQUEST);
-        }
-
-        cartService.addCart(member.getId(), cartForm.getItemId(), cartForm.getCount());
-        return ResponseEntity.ok("success");
-    }
-
-    @DeleteMapping("/cart")
-    @ResponseBody
-    public ResponseEntity<String> deleteCartItem(@RequestBody CartItemForm form) {
-
-//        log.info("itemId={}", form.getCartItemId());
-
-        if (cartService.findCartItem(form.getCartItemId()) == null) {
-            return new ResponseEntity<String>("다시 시도해주세요.", HttpStatus.NOT_FOUND);
-        }
-
-        cartService.deleteCartItem(form.getCartItemId());
-
-        return ResponseEntity.ok("success");
-    }
-
-    //다른 컨트롤러에서도 사용하기 위해 static
-    public static Member getMember(HttpServletRequest request) {
-
-        HttpSession session = request.getSession(false);
-
-        //비로그인 사용자
-        if (session == null || session.getAttribute(SessionConst.LOGIN_MEMBER) == null) {
-            return null;
-        }
-
-        //세션에 저장되어있는 회원정보 가져오기
-        Member member = (Member) session.getAttribute(SessionConst.LOGIN_MEMBER);
-        return member;
     }
 
 }
