+++ backend/src/main/java/com/dsapkl/backend/service/CustomUserDetailsService.java
@@ -12,7 +12,6 @@
 //@Service
 //@RequiredArgsConstructor
 //public class CustomUserDetailsService implements UserDetailsService {
-//
 //    private final MemberRepository memberRepository;
 //
 //    @Override
