+++ backend/src/main/java/com/dsapkl/backend/util/SessionConst.java
@@ -1,4 +1,4 @@
-package com.dsapkl.backend.controller;
+package com.dsapkl.backend.util;
 
 public class SessionConst {
     public static final String LOGIN_MEMBER ="loginMember";
